chore(tests): Improve code maintainability  - Refactors tests files to improve maintainability  - Sets codeclimate similar-code threshold to 50 [fixes #161471006]
